pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
extra_standard_library = "setuptools,typing_extensions"
force_grid_wrap = 0
include_trailing_comma = true
line_length = 100
multi_line_output = 3
skip = ["/home/circleci/repo/venv", "./venv", "./.venv"]
use_parentheses = true

[tool.mypy]
allow_redefinition = false
check_untyped_defs = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = ".:test:content_aware_timelapse"
no_implicit_optional = true
scripts_are_modules = true
strict_optional = false
warn_incomplete_stub = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Devon Bray"]
description = "Create timelapses from video that change speed based on the content found in the video. Boring sections are sped over."
name = "content_aware_timelapse"
readme = "README.md"
version = "0.3.0"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.4.1"
torchvision = "^0.19.1"
opencv-python = "^4.10.0.84"
ffmpeg-python = "^0.2.0"
vidgear = "^0.3.3"
more-itertools = "^10.5.0"
click = "^8.1.7"
tqdm = "^4.66.5"
timm = "^1.0.15"
pillow = "^10.4.0"
h5py = "^3.12.1"
sentinels = "^1.0.0"
joblib = "^1.4.2"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
pkginfo = "^1.12.1.2"
clip = {git = "https://github.com/openai/CLIP.git"}
ftfy = "^6.3.1"
regex = "^2024.11.6"
tabulate = "^0.9.0"
matplotlib = "^3.10.3"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pylint = "*"
pydevd-pycharm = ">=231.8109.197,<231.8110.0"
wheel = "^0.38"

[tool.poetry.group.test.dependencies]
coverage = "*"
pytest = "*"
pytest-cov = "*"

[tool.pytest.ini_options]
addopts = "-rsxX --showlocals --strict-markers"

[tool.pytest.ini_options.markers]
devtest = "temporary development mark that can be used as needed (should not be merged to master)"
integration = "tests that access outside services (s3, APIs), don't run in CircleCI"
timeout = "timeout for test in seconds, from `pytest-timeout` module"
